[#897] Geometry Matrix clean-up: moved _compute_los() and _compute_vos() to dedicated _class09_compute_broadband.py
